+++ server/src/sessions/api.go
@@ -2,26 +2,21 @@ package sessions
 
 import (
 	"context"
-	"net/url"
 
 	"github.com/google/uuid"
 )
 
-type SessionService interface {
-	Create(ctx context.Context, boardID, userID uuid.UUID) (*BoardSession, error)
-	Update(ctx context.Context, body BoardSessionUpdateRequest) (*BoardSession, error)
-	UpdateAll(ctx context.Context, body BoardSessionsUpdateRequest) ([]*BoardSession, error)
-	UpdateUserBoards(ctx context.Context, body BoardSessionUpdateRequest) ([]*BoardSession, error)
-	Get(ctx context.Context, boardID, userID uuid.UUID) (*BoardSession, error)
-	GetAll(ctx context.Context, boardID uuid.UUID, filter BoardSessionFilter) ([]*BoardSession, error)
-	GetUserConnectedBoards(ctx context.Context, user uuid.UUID) ([]*BoardSession, error)
+type UserService interface {
+	CreateAnonymous(ctx context.Context, name string) (*User, error)
+	CreateAppleUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	CreateAzureAdUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	CreateGitHubUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	CreateGoogleUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	CreateMicrosoftUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	CreateOIDCUser(ctx context.Context, id, name, avatarUrl string) (*User, error)
+	Update(ctx context.Context, body UserUpdateRequest) (*User, error)
+	Get(ctx context.Context, id uuid.UUID) (*User, error)
 
-	Connect(ctx context.Context, boardID, userID uuid.UUID) error
-	Disconnect(ctx context.Context, boardID, userID uuid.UUID) error
-
-	Exists(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
-	ModeratorSessionExists(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
-	IsParticipantBanned(ctx context.Context, boardID, userID uuid.UUID) (bool, error)
-
-	BoardSessionFilterTypeFromQueryString(query url.Values) BoardSessionFilter
+	IsUserAvailableForKeyMigration(ctx context.Context, id uuid.UUID) (bool, error)
+	SetKeyMigration(ctx context.Context, id uuid.UUID) (*User, error)
 }
